Footers: Extract shared components
Since we are exporting *, we get some errors as they share some
exported components. Let's just go ahead and pull those out into
another file.
